Moved to main workspace

+++ conf/patches/me.java
@@ -5,6 +5,6 @@ public class me
     public static int c(double d)
     {
         int i = (int)d;
-        return d >= (double)i ? i : i - 1;
+        return d >= i ? i : i - 1;
     }
 }
